+++ src/operation/iRCX/source/module/spef_writer/sw_data_manager/SPEFNameMap.hpp
@@ -25,25 +25,25 @@ class SPEFNameMap
   SPEFNameMap() = default;
   ~SPEFNameMap() = default;
   // getter
-  std::unordered_map<std::string, size_t>& get_net_name_to_id_map() { return _net_name_to_id_map; }
-  std::unordered_map<std::string, size_t>& get_port_name_to_id_map() { return _port_name_to_id_map; }
-  std::unordered_map<std::string, size_t>& get_instance_name_to_id_map() { return _instance_name_to_id_map; }
-  std::map<size_t, std::string>& get_id_to_net_name_map() { return _id_to_net_name_map; }
-  std::map<size_t, std::string>& get_id_to_port_name_map() { return _id_to_port_name_map; }
-  std::map<size_t, std::string>& get_id_to_instance_name_map() { return _id_to_instance_name_map; }
-  size_t get_next_id() const { return _next_id; }
+  std::unordered_map<std::string, int32_t>& get_net_name_to_id_map() { return _net_name_to_id_map; }
+  std::unordered_map<std::string, int32_t>& get_port_name_to_id_map() { return _port_name_to_id_map; }
+  std::unordered_map<std::string, int32_t>& get_instance_name_to_id_map() { return _instance_name_to_id_map; }
+  std::map<int32_t, std::string>& get_id_to_net_name_map() { return _id_to_net_name_map; }
+  std::map<int32_t, std::string>& get_id_to_port_name_map() { return _id_to_port_name_map; }
+  std::map<int32_t, std::string>& get_id_to_instance_name_map() { return _id_to_instance_name_map; }
+  int32_t get_next_id() const { return _next_id; }
   // setter
-  void set_next_id(size_t next_id) { _next_id = next_id; }
+  void set_next_id(int32_t next_id) { _next_id = next_id; }
   // function
 
  private:
-  std::unordered_map<std::string, size_t> _net_name_to_id_map;
-  std::unordered_map<std::string, size_t> _port_name_to_id_map;
-  std::unordered_map<std::string, size_t> _instance_name_to_id_map;
-  std::map<size_t, std::string> _id_to_net_name_map;
-  std::map<size_t, std::string> _id_to_port_name_map;
-  std::map<size_t, std::string> _id_to_instance_name_map;
-  size_t _next_id = 1;
+  std::unordered_map<std::string, int32_t> _net_name_to_id_map;
+  std::unordered_map<std::string, int32_t> _port_name_to_id_map;
+  std::unordered_map<std::string, int32_t> _instance_name_to_id_map;
+  std::map<int32_t, std::string> _id_to_net_name_map;
+  std::map<int32_t, std::string> _id_to_port_name_map;
+  std::map<int32_t, std::string> _id_to_instance_name_map;
+  int32_t _next_id = 1;
 };
 
 }  // namespace ircx
